requirements.txt
black
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
loguru
pytest
fastapi[all]
neuralprophet
scikit-learn
optuna
influxdb-client[ciso]
python-dotenv
requests
apscheduler